labs-server-xs22/docker/script_init/requirements.txt
urllib3
python-dotenv
pathlib
mysql.connector
